refactor: clean up ugly code
