+++ packages/extension/src/ui/App.tsx
@@ -1,6 +1,6 @@
 import { ThemeProvider, createTheme } from "@mui/material"
 import { FC, Suspense } from "react"
-import { Navigate, Outlet, Route, Routes, useNavigate } from "react-router-dom"
+import { Navigate, Outlet, Route, Routes } from "react-router-dom"
 import styled, { createGlobalStyle } from "styled-components"
 import { normalize } from "styled-normalize"
 import { SWRConfig } from "swr"
@@ -26,6 +26,7 @@ import { LockScreen } from "./screens/LockScreen"
 import { NewWalletScreen } from "./screens/NewWalletScreen"
 import { NftScreen } from "./screens/NftScreen"
 import { ResetScreen } from "./screens/ResetScreen"
+import { SeedRecoveryPasswordScreen } from "./screens/SeedRecoveryPasswordScreen"
 import { SeedRecoveryScreen } from "./screens/SeedRecoveryScreen"
 import { SettingsDappConnectionsScreen } from "./screens/SettingsDappConnectionsScreen"
 import { SettingsNetworkFormScreen } from "./screens/SettingsNetworkForm"
@@ -38,15 +39,7 @@ import { UpgradeScreen } from "./screens/UpgradeScreen"
 import { WelcomeScreen } from "./screens/WelcomeScreen"
 import { useActions, useActionsSubscription } from "./states/actions"
 import { useAppState } from "./states/app"
-import { useBackupRequired } from "./states/backupDownload"
-import {
-  useSeedRecover,
-  validateAndSetPassword,
-  validateSeedRecoverStateIsComplete,
-} from "./states/seedRecover"
 import { useSelectedNetwork } from "./states/selectedNetwork"
-import { recoverBySeedPhrase } from "./utils/messaging"
-import { recover } from "./utils/recovery"
 import { swrCacheProvider } from "./utils/swrCache"
 
 const GlobalStyle = createGlobalStyle`
@@ -118,10 +111,7 @@ const Screen: FC = () => {
 
   const { isLoading } = useAppState()
   const { actions } = useActions()
-
   const [selectedCustomNetwork] = useSelectedNetwork()
-
-  const navigate = useNavigate()
 
   if (isLoading) {
     return <LoadingScreen />
@@ -146,25 +136,7 @@ const Screen: FC = () => {
         <Route path={routes.seedRecovery()} element={<SeedRecoveryScreen />} />
         <Route
           path={routes.seedRecoveryPassword()}
-          element={
-            <NewWalletScreen
-              overrideTitle="New password"
-              overrideSubmitText="Continue"
-              overrideSubmit={async ({ password }) => {
-                try {
-                  validateAndSetPassword(password)
-                  const state = useSeedRecover.getState()
-                  if (validateSeedRecoverStateIsComplete(state)) {
-                    await recoverBySeedPhrase(state.seedPhrase, state.password)
-                    useBackupRequired.setState({ isBackupRequired: false }) // as the user recovered their seed, we can assume they have a backup
-                    navigate(await recover())
-                  }
-                } catch {
-                  console.error("seed phrase is invalid")
-                }
-              }}
-            />
-          }
+          element={<SeedRecoveryPasswordScreen />}
         />
         <Route path={routes.lockScreen()} element={<LockScreen />} />
         <Route path={routes.reset()} element={<ResetScreen />} />
